+++ bin/test.sh
@@ -14,9 +14,10 @@ echo "##################################################" >> ../onlysnarf/logs/t
 # onlysnarfpy -debug -verbose -verbose -verbose -browser remote-firefox -prefer-local -remote-host 47.156.172.175 -remote-port 8888 -source google -category video -keep -show -login google -skip-download
 # onlysnarfpy -debug -verbose -verbose -verbose -browser auto-remote -source google -category video -keep -skip-download -login google -remote-host 47.156.172.175 -remote-port 8888
 
+onlysnarfpy -debug -verbose -verbose -verbose -source local -category video
+# onlysnarfpy -debug -verbose -verbose -verbose -source remote -remote-username skeetzo
 
-
-onlysnarf -debug -verbose -verbose -verbose -browser firefox -prefer-local -remote-host 192.168.1.66 -login onlyfans
+# onlysnarf -debug -verbose -verbose -verbose
 
 # onlysnarfpy -action test -debug -verbose -verbose -verbose -browser auto-remote -keep -remote-host 47.156.172.175 -remote-port 8888 -login google